codegen: some object api progress
add a common.PackOptions param to unpack() and unpackTo() methods which
allows the user to provide an allocator that is used to create pointer
types (structs/vectors).

this patch allows all_tests.zig to compile and run but shows an error
in code generated by structBuilderBody(). the field ordering in
generated Create() is different from the go code.